BF - refresh contrast when adding from design
There was a serious bug on adding contrasts, for F contrasts, because
the F contrasts in SPM carry with them some sub-fields that have been
produced from the filtered design matrix.  Marsbar is almost certain to
re-estimate this matrix, and so the sub-fields no longer match the
actual filtered design, and invalid F values result.