MISC - allow for SPM8 t statistic fudge factor
SPM8 adds a fudge factor to the standard error for t statistics so there is a
small difference between the marsbar and SPM t statistic.  Check for this
difference when running the test rig and allow test rig to pass when the
difference is probably due to the fudge factor.