RF - only refresh contrasts for estimated designs
The marsbar design interface is useful for SPM designs, but refreshing the
contrasts on loading SPM designs wipes out the contrast images.  As the problem
is only for making statistics from the contrasts, and statistics require
MarsBaR estimated designs, it is enough to refresh MarsBaR estimated designs.

This allows the tutorial to work because it depended on getting contrast image
filenames out of an SPM design loaded in MarsBaR.